+++ src/main/java/ro/webdata/normalization/timespan/ro/regex/date/DateRegex.java
@@ -12,42 +12,42 @@ public class DateRegex {
     // d{3,} allows avoiding the month-day pattern (E.g.: "noiembrie 22")
     private static final String DATE_DMY_DOT =
             "("
-                + "\\d{1,2}[\\.]{1}\\d{2}[\\.]{1}\\d{3,}"
+                + "\\d{1,2}\\.\\d{1,2}\\.\\d{3,}"
                 + AD_BC_OPTIONAL
             + ")";        // E.g.: "01.01.1911"
     private static final String DATE_DMY_DOT_PARTIAL =
             "("
-                + "\\d{1,2}[\\.]{1}\\d{2}"
+                + "\\d{1,2}\\.\\d{2}"
                 + "("
-                    + "[\\.]{1}\\d{3,}"
+                    + "\\.\\d{3,}"
                 + ")*"
                 + AD_BC_OPTIONAL
             + ")";
 
     private static final String DATE_DMY_SLASH =
             "("
-                + "\\d{1,2}[/]{1}\\d{2}[/]{1}\\d{3,}"
+                + "\\d{1,2}\\/\\d{2}\\/\\d{3,}"
                 + AD_BC_OPTIONAL
             + ")";          // E.g.: "21/01/1916"
     private static final String DATE_DMY_SLASH_PARTIAL =
             "("
-                + "\\d{1,2}[/]{1}\\d{2}"
+                + "\\d{1,2}\\/\\d{2}"
                 + "("
-                    + "[/]{1}\\d{3,}"
+                    + "\\/\\d{3,}"
                 + ")*"
                 + AD_BC_OPTIONAL
             + ")";
 
     private static final String DATE_DMY_SPACE =
             "("
-            + "\\d{1,2}\\s*{1}\\d{2}\\s*{1}\\d{3,}"
+            + "\\d{1,2}\\s*\\d{2}\\s*\\d{3,}"
             + AD_BC_OPTIONAL
             + ")";          // E.g.: "7 06 1911"
     private static final String DATE_DMY_SPACE_PARTIAL =
             "("
-                + "\\d{1,2}\\s*{1}\\d{2}"
+                + "\\d{1,2}\\s*\\d{2}"
                 + "("
-                    + "\\s*{1}\\d{3,}"
+                    + "\\s*\\d{3,}"
                 + ")*"
                 + AD_BC_OPTIONAL
             + ")";
@@ -55,14 +55,14 @@ public class DateRegex {
     private static final String DATE_DMY_TEXT =
             "("
                 + "\\d{1,2}[,\\s]+"
-                + MONTHS_RO
+                + MONTHS
                 + "[,\\s]+\\d{3,}"
                 + AD_BC_OPTIONAL
             + ")";  // E.g.: "9 iulie 1807"
     public static final String DATE_DMY_TEXT_PARTIAL =
             "("
                 + "\\d{1,2}[, ]+"
-                + MONTHS_RO
+                + MONTHS
                 + "("
                     + "[,\\s]+\\d{3,}"
                     + AD_BC_OPTIONAL
@@ -78,35 +78,36 @@ public class DateRegex {
     private static final String DATE_YMD_TEXT =
             "("
                 + "\\d{3,}[,\\s]+"
-                + MONTHS_RO
+                + MONTHS
                 + "[,\\s]+\\d{1,2}"
                 + AD_BC_OPTIONAL
             + ")";  // E.g.: "1752 aprilie 25"
 
-    private static final String DATE_DMY_INTERVAL_START =
-            "("
-                + "(" + TEXT_START + DATE_DMY_DOT_PARTIAL + ")" + REGEX_OR
-                + "(" + TEXT_START + DATE_DMY_SLASH_PARTIAL + ")" + REGEX_OR
-                + "(" + TEXT_START + DATE_DMY_SPACE_PARTIAL + ")" + REGEX_OR
-                + "(" + TEXT_START + DATE_DMY_TEXT_PARTIAL + ")"
+    private static final String DATE_DMY_INTERVAL_START = TEXT_START
+            + "("
+                + DATE_DMY_DOT_PARTIAL + REGEX_OR
+                + DATE_DMY_SLASH_PARTIAL + REGEX_OR
+                + DATE_DMY_SPACE_PARTIAL + REGEX_OR
+                + DATE_DMY_TEXT_PARTIAL + REGEX_OR
+                + "\\d{1,2}\\s+" + MONTHS + "?"
             + ")";
-    private static final String DATE_YMD_INTERVAL_START =
-            "("
-                + "(" + TEXT_START + DATE_YMD_DASH + ")" + REGEX_OR
-                + "(" + TEXT_START + DATE_YMD_TEXT + ")"
+    private static final String DATE_YMD_INTERVAL_START = TEXT_START
+            + "("
+                + DATE_YMD_DASH + REGEX_OR
+                + DATE_YMD_TEXT
             + ")";
     private static final String DATE_DMY_INTERVAL_END =
             "("
-                + "(" + DATE_DMY_DOT + TEXT_END + ")" + REGEX_OR
-                + "(" + DATE_DMY_SLASH + TEXT_END + ")" + REGEX_OR
-                + "(" + DATE_DMY_SPACE + TEXT_END + ")" + REGEX_OR
-                + "(" + DATE_DMY_TEXT + TEXT_END + ")"
-            + ")";
+                + DATE_DMY_DOT + REGEX_OR
+                + DATE_DMY_SLASH + REGEX_OR
+                + DATE_DMY_SPACE + REGEX_OR
+                + DATE_DMY_TEXT
+            + ")" + TEXT_END;
     private static final String DATE_YMD_INTERVAL_END =
             "("
-                + "(" + DATE_YMD_DASH + TEXT_END + ")" + REGEX_OR
-                + "(" + DATE_YMD_TEXT + TEXT_END + ")"
-            + ")";
+                + DATE_YMD_DASH + REGEX_OR
+                + DATE_YMD_TEXT
+            + ")" + TEXT_END;
 
     public static final String DATE_DMY_INTERVAL = CASE_INSENSITIVE
             + "("
@@ -121,16 +122,16 @@ public class DateRegex {
                 + DATE_YMD_INTERVAL_END
             + ")";
 
-    public static final String DATE_DMY_OPTIONS = CASE_INSENSITIVE
+    public static final String DATE_DMY_OPTIONS = CASE_INSENSITIVE + TEXT_START
             + "("
-                + "(" + TEXT_START + DATE_DMY_DOT + TEXT_END + ")" + REGEX_OR
-                + "(" + TEXT_START + DATE_DMY_SLASH + TEXT_END + ")" + REGEX_OR
-                + "(" + TEXT_START + DATE_DMY_SPACE+ ")" + REGEX_OR
-                + "(" + TEXT_START + DATE_DMY_TEXT + TEXT_END + ")"
-            + ")";
-    public static final String DATE_YMD_OPTIONS = CASE_INSENSITIVE
+                + DATE_DMY_DOT + REGEX_OR
+                + DATE_DMY_SLASH + REGEX_OR
+                + DATE_DMY_SPACE + REGEX_OR
+                + DATE_DMY_TEXT
+            + ")" + TEXT_END;
+    public static final String DATE_YMD_OPTIONS = CASE_INSENSITIVE + TEXT_START
             + "("
-                + "(" + TEXT_START + DATE_YMD_DASH + TEXT_END + ")" + REGEX_OR
-                + "(" + TEXT_START + DATE_YMD_TEXT + TEXT_END + ")"
-            + ")";
+                + DATE_YMD_DASH + REGEX_OR
+                + DATE_YMD_TEXT
+            + ")" + TEXT_END;
 }